moduł ektraktora cech

+++ src/main/java/pl/mowk/ksr/extractions/NumberOfWords.java
@@ -3,7 +3,7 @@ package pl.mowk.ksr.extractions;
 import pl.mowk.ksr.data.Article;
 
 public class NumberOfWords {
-    static public int count(Article art){
+    static public double count(Article art){
         String[] allWords = art.getBody().split(" ");
         int wordCount = 0;
         for (String i : allWords){
